pyproject.toml
[project]
name = "bee-thresholding-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "opencv-python>=4.11.0.86",
  "pandas>=2.2.3",
  "pillow>=11.2.1",
  "scikit-learn>=1.6.1",
  "tqdm>=4.67.1",
]
